reworked simulation backends in prep of move at eof rework
