+++ bsb/simulators/arbor/__init__.py
@@ -1,3 +1,6 @@
+from bsb.simulation import SimulationBackendPlugin
+from .simulation import ArborSimulation
 from .adapter import ArborAdapter
 
-__plugin__ = ArborAdapter
+
+__plugin__ = SimulationBackendPlugin(Simulation=ArborSimulation, Adapter=ArborAdapter)
