+++ bsb/simulators/nest/__init__.py
@@ -1,0 +1,6 @@
+from bsb.simulation import SimulationBackendPlugin
+from .adapter import NestAdapter
+from .simulation import NestSimulation
+
+
+__plugin__ = SimulationBackendPlugin(Simulation=NestSimulation, Adapter=NestAdapter)
