+++ bsb/simulators/neuron/__init__.py
@@ -1,3 +1,6 @@
+from bsb.simulation import SimulationBackendPlugin
 from .adapter import NeuronAdapter
+from .simulation import NeuronSimulation
 
-__plugin__ = NeuronAdapter
+
+__plugin__ = SimulationBackendPlugin(Simulation=NeuronSimulation, Adapter=NeuronAdapter)
