+++ bsb/simulators/neuron/devices/spike_generator.py
@@ -1,5 +1,5 @@
 from .... import config
-from ..adapter import NeuronDevice
+from ..device import NeuronDevice
 from ....simulation.results import SimulationRecorder, PresetPathMixin, PresetMetaMixin
 from ....exceptions import *
 from ....reporting import report, warn
